refactor upgrades tests
