+++ .github/workflows/e2e-upgrade.yaml
@@ -12,6 +12,7 @@ jobs:
       matrix:
         include:
           - test: TestV4ToV5ChainUpgrade
+          - test: TestV5ToV6ChainUpgrade
     steps:
       - uses: actions/checkout@v3
       - uses: actions/setup-go@v3
